steam login fixes

+++ XIV on Mac/Steamworks/Steamworks.swift
@@ -37,7 +37,7 @@ struct Steam {
         let ticketString = rawTicketSteam.map { String(format: "%02hhx", $0) }.joined()
         let rawTicket = ticketString.compactMap { $0.asciiValue } + [0]
         let ticketSum = rawTicket.map { UInt16($0) }.reduce(0, &+)
-        let ticketSumTruncated = UInt8(truncating: NSNumber(value: ticketSum))
+        let ticketSumTruncated = Int16(truncating: NSNumber(value: ticketSum))
         let time = 60 * ((steamworks.serverRealTime - 5) / 60);
         let rand = CrtRand(seed: time ^ UInt32(ticketSumTruncated))
         let blowfishKey = String(format: "%08x#un@e=x>", time)
@@ -75,7 +75,7 @@ extension Data {
       return UInt32(littleEndian: intBits)
     }
     
-    func squareBase64EncodedString() -> String{
+    func squareBase64EncodedString() -> String {
         self.base64EncodedString()
             .replacingOccurrences(of: "+", with: "-")
             .replacingOccurrences(of: "/", with: "_")
